Update configuration.rst
I try to used the start.sh script proposed in this file, but it don't work on my computer (the APP_CONFIG_FILE envar was not found). After searching, I saw that I needed to export the envvar in start.sh. I don't know if it's an error, so I prefer to make a pull request if it's one.